Launcher3: Move restart check to home page
It's just a flag check. And avoids having to implement a listener for the home key (and still miss home gesture).
Also, restarting the launcher when navigating to previous settings page doesn't seem right.
This way, the launcher is restarted once, if needed, when user goes back to home page, which is where the changes are visible, anyway.

Change-Id: I80b7541cb4e4881d1136f05018cdee10e40eff5d
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>